+++ src/NetBuilder/GaussMethod.cc
@@ -181,8 +181,8 @@ void add_line(GeneratingMatrix& T, GeneratingMatrix& D, std::map<int, int>& Orig
 }
 
 int iteration_on_k(std::vector<GeneratingMatrix>& Origin_Mats, int k, bool verbose=false){
-    int s = Origin_Mats.size();
-    int m = Origin_Mats[0].nCols();
+    int s = (int) Origin_Mats.size();
+    int m = (int) Origin_Mats[0].nCols();
 
     std::vector<std::vector<int>> lines_order = compositions(k, s);
     
@@ -260,8 +260,8 @@ int iteration_on_k(std::vector<GeneratingMatrix>& Origin_Mats, int k, bool verbo
 
 int GaussMethod::computeTValue(std::vector<GeneratingMatrix> Origin_Mats, int maxSubProj, int verbose=0)
 {
-    int m = Origin_Mats[0].nRows();
-    int s = Origin_Mats.size();
+    int m = (int) Origin_Mats[0].nRows();
+    int s = (int) Origin_Mats.size();
     if (s == 1){    // to be modified!
         return 0;
     }
